requirements.txt
pillow>=8.2.0
discord>=1.0.1
discord.py>=1.7.2
async-timeout>=3.0.1
requests>=2.25.0
youtube-dl>=2021.5.16

